(TODO message) Store all annotations in Kotlin metadata on JVM
Annotations are declared in the common metadata.proto, but at the
moment, are only serialized on JVM. There are proto extensions for
annotations in klib, they will eventually be migrated to annotations in
the common proto (TODO: KT-XXXXX).

At the moment, the stored annotations are not used anywhere in the
compiler. The only place where they're used is protoDifferenceUtils,
which makes incremental compilation correctly track difference in
annotations. They will also be used in a subsequent commit in
kotlinx-metadata-jvm, and tests on kotlinp.

 #KT-57919 Fixed