[Fir2IR] Build fake overrides after actualization of classes
To build fake overrides correctly, we need to do it, when
there are no actual supertypes.

On the other hand, we can't do it after actualizing callables,
because some callable can be actualized to fake override.

That's why we need build fake overrides during actualization,
when classes are already actualized, while callables are not.

^KT-63644