[K/N][test] Support for running Kotlin/Native tests using XCTest
This change adds a library with cinterop that has XCTest wrapper around
Kotlin/Native tests (that are kotlin.test.Test methods).
This library should be compiled with test code using the option
`-produce test_bundle`.

The basic idea is to make XCTest be able to resolve separate test cases
and correctly show them in test reports. This was achieved by wrapping
test cases with dynamically created invocation methods. Test listeners
are integrated with XCTest observation to make it possible to have the
same ability to report with GTest or TeamCity logging.

Gradle build files use MPP Gradle plugin and use a bootstrap version
of K/N. Property `kotlin.native.home` was moved to the kotlin-native
subproject to not override this project's K/N distribution, that is
being controlled in the KGP with the same property.

Part of ^KT-58928