kotlinx-metadata-klib: rename annotations -> klibAnnotations
In a subsequent commit, annotations will be added as members for
KmClass, KmFunction, etc., so these extensions will be shadowed. In the
future, if/when we start to use the common annotations in klib metadata
instead of the extensions, we will remove these extension properties.
Before that happens though, `annotations` for klib metadata will be
empty, and the actual annotations will be in `klibAnnotations`.