[stdlib] Don't pass 'stdlib' twice to Kotlin/JS & Kotlin/WASM compilers
The KLIB resolver is a component inside the compiler that manipulates
dependency libraries before actually reading them. It is a source
of numerous problems (such as, KT-63573) and is going to be
gradually deprecated and removed in the future (see KFC-737).

We are going to relax restrictions inside the KLIB resolver, so that
libraries that happen to have the same `unique_name` in manifest
would be all passed further as opposed to selecting just one among them
as it works right now.

However, we would need first to make some preparation in order to
avoid reading the same library multiple times in the compiler.